added method and tests for zoom level method
