+++ lib/zoom_level_calculator.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/services.dart';
+
+Future<double> getZoomLevel(double viewWidthInLogicalPixels, double documentWidthInMM, {double scaleOfDisplayedObject = 1}) async {
+  double dpm = await MethodChannel("measurements").invokeMethod("getPhysicalPixelsPerMM");
+
+  double screenWidth = viewWidthInLogicalPixels / dpm;
+
+  return documentWidthInMM / (screenWidth * scaleOfDisplayedObject);
+}
