+++ test/measurement_bloc_test.dart
@@ -8,12 +8,6 @@ import 'package:measurements/bloc/measurement_bloc.dart';
 void main() {
   const zoomLevel = 1.0;
   const scale = 1.0;
-  const viewWidth = 400.0;
-  const dpm = 20.0;
-
-  const expectedZoomFactor = 10.0;
-
-  const MethodChannel channel = MethodChannel('measurements');
 
   Function(List<double>) distanceCallback;
   List<double> actualDistances = List();
@@ -23,14 +17,6 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   setUpAll(() {
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      if (methodCall.method == "getPhysicalPixelsPerMM") {
-        return dpm;
-      } else {
-        return -1;
-      }
-    });
-
     distanceCallback = (List<double> distances) {
       actualDistances = distances;
       print("Updated points $distances");
@@ -40,7 +26,6 @@ void main() {
   setUp(() {
     classUnderTest = MeasurementBloc(Size(200, 300), distanceCallback);
 
-    classUnderTest.viewWidth = viewWidth;
     classUnderTest.scale = scale;
     classUnderTest.zoomLevel = zoomLevel;
     classUnderTest.measuring = true;
@@ -48,16 +33,6 @@ void main() {
 
   tearDown(() {
     classUnderTest.dispose();
-  });
-
-  tearDownAll(() {
-    channel.setMockMethodCallHandler(null);
-  });
-
-  test("setZoomToOriginalSize", () async {
-    double zoomFactor = await classUnderTest.getZoomFactorForOriginalSize();
-
-    expect(zoomFactor, expectedZoomFactor);
   });
 
   test("getDistanceFromHorizontalPoints", () async {
